linalg: lower aten.linalg.vector_norm to linalg on tensors
This patch extends the conversion of reduction operators to support
lowering of AtenLinalgVectorNormOp, in addition to adding a handful of
end-to-end tests to validate the lowering.

There exist several opportunities to make this lowering optimal and
robust.  For instance, in its current form, the translation does not
support L0 norms.  For L1 norms, we don't need to raise each element to
the power 1.0.  Similarly, L2 norms could benefit from strength
reduction.  Sadly, the canonicalization pass is not able to apply these
optimizations.